backend/foodgram/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==23.3
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
Pillow==9.0.0
python-dotenv==1.0.0